requirements.txt
apache-airflow>=2.8.0
dbt-duckdb>=1.7.0
duckdb>=0.10.0
requests>=2.31.0
python-dotenv>=1.0.0
pendulum>=3.0.0
